commited

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -11,7 +11,7 @@
 hash_table_t *hash_table_create(unsigned long int size)
 {
 	hash_table_t *t;
-	unsigned int i;
+	unsigned long int i;
 
 	t = malloc(sizeof(hash_table_t));
 	if (t == NULL)
